backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
pydantic
python-multipart
boto3
requests
openai
PyMuPDF
google-generativeai
aiofiles  
email-validator